requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
psycopg2==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.0.1
pyzmq==26.2.0
requests==2.32.3
rsa==4.9
Telethon==1.38.1
urllib3==2.3.0
